pyproject.toml
[project]
name = "investment-rag-bot"
version = "0.1.0"
description = "Investment Banking RAG Bot with Langchain, Pinecone and Google Gemini"
authors = [
    {name = "vansh-561",email = "vanshagrawal561@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "pypdf (==5.5.0)",
    "langchain (>=0.1.4)",
    "langchain-community (>=0.3.23)",
    "langchain-core (>=0.1.22)",
    "langchain-google-genai (>=0.0.6)",
    "google-generativeai (>=0.4.0)",
    "pydantic (>=2.0.0)",
    "watchdog (>=3.0.0)",
    "streamlit (>=1.45.0,<2.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "tqdm (==4.67.1)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pinecone (>=6.0.2,<7.0.0)"
]

[tool.poetry]
packages = [{include = "investment_rag_bot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
